# services/frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Smart Parks Protect",
  "version": "committed"
 },
 "paths": {
...
  "/api/v1/projects/{project_id}/analytics/series": {
   "get": {
    "tags": [
     "analytics"
    ],
    "summary": "Series",
    "description": "Bucketed aggregates of measurements attributed to the project. One series per metric and\nentity (or device). Boolean metrics aggregate as 0 and 1, so `mean` is the fraction true.",
    "operationId": "series_api_v1_projects__project_id__analytics_series_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Project Id"
      }
     },
     {
      "name": "metric",
      "in": "query",
      "required": true,
      "schema": {
       "type": "array",
       "items": {
        "type": "string"
       },
       "minItems": 1,
       "maxItems": 20,
       "title": "Metric"
      }
     },
     {
      "name": "entity_id",
      "in": "query",
      "required": false,
      "schema": {
       "type": "array",
       "items": {
        "type": "string",
        "format": "uuid"
       },
       "maxItems": 20,
       "title": "Entity Id"
      }
     },
     {
      "name": "device_id",
      "in": "query",
      "required": false,
      "schema": {
       "type": "array",
       "items": {
        "type": "string",
        "format": "uuid"
       },
       "maxItems": 20,
       "title": "Device Id"
      }
     },
     {
      "name": "data_source_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Data Source Id"
      }
     },
     {
      "name": "from",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date-time"
        },
        {
         "type": "null"
        }
       ],
       "title": "From"
      }
     },
     {
      "name": "to",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date-time"
        },
        {
         "type": "null"
        }
       ],
       "title": "To"
      }
     },
     {
      "name": "bucket",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "One of 1s, 10s, 1m, 5m, 15m, 1h, 6h, 1d, 7d or `all`; default automatic",
       "title": "Bucket"
      },
      "description": "One of 1s, 10s, 1m, 5m, 15m, 1h, 6h, 1d, 7d or `all`; default automatic"
     },
     {
      "name": "agg",
      "in": "query",
      "required": false,
      "schema": {
       "type": "array",
       "items": {
        "$ref": "#/components/schemas/Aggregate"
       },
       "title": "Agg"
      }
     },
     {
      "name": "group_by",
      "in": "query",
      "required": false,
      "schema": {
       "$ref": "#/components/schemas/GroupBy",
       "default": "entity"
      }
     },
     {
      "name": "layout",
      "in": "query",
      "required": false,
      "schema": {
       "$ref": "#/components/schemas/Layout",
       "default": "series"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SeriesResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/projects/{project_id}/analytics/rows": {
   "get": {
    "tags": [
     "analytics"
    ],
    "summary": "Rows",
    "description": "Normalized measurement rows behind a bucket: the drill-down from an aggregate. Each row\ncarries its source event and trace, so the next step down is the source event detail.",
    "operationId": "rows_api_v1_projects__project_id__analytics_rows_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Project Id"
      }
     },
     {
      "name": "metric",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Metric"
      }
     },
     {
      "name": "entity_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Entity Id"
      }
     },
     {
      "name": "device_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Device Id"
      }
     },
     {
      "name": "data_source_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Data Source Id"
      }
     },
     {
      "name": "from",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date-time"
        },
        {
         "type": "null"
        }
       ],
       "title": "From"
      }
     },
     {
      "name": "to",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date-time"
        },
        {
         "type": "null"
        }
       ],
       "title": "To"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "cursor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "key of the last item of the previous page",
       "title": "Cursor"
      },
      "description": "key of the last item of the previous page"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PageResponse_MeasurementRow_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/projects/{project_id}/analytics/saved-views": {
   "get": {
    "tags": [
     "analytics"
    ],
    "summary": "List Saved Views",
    "description": "Data Explorer configurations shared by the project (decision D42).",
    "operationId": "list_saved_views_api_v1_projects__project_id__analytics_saved_views_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Project Id"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "cursor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "key of the last item of the previous page",
       "title": "Cursor"
      },
      "description": "key of the last item of the previous page"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PageResponse_SavedViewRead_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "analytics"
    ],
    "summary": "Create Saved View",
    "description": "Any member can save a view; names are unique per project.",
    "operationId": "create_saved_view_api_v1_projects__project_id__analytics_saved_views_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Project Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SavedViewCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SavedViewRead"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/projects/{project_id}/analytics/saved-views/{view_id}": {
   "patch": {
    "tags": [
     "analytics"
    ],
    "summary": "Update Saved View",
    "operationId": "update_saved_view_api_v1_projects__project_id__analytics_saved_views__view_id__patch",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "view_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "View Id"
      }
     },
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Project Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SavedViewUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SavedViewRead"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "analytics"
    ],
    "summary": "Delete Saved View",
    "operationId": "delete_saved_view_api_v1_projects__project_id__analytics_saved_views__view_id__delete",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "view_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "View Id"
      }
     },
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Project Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "Aggregate": {
    "type": "string",
    "enum": [
     "mean",
     "min",
     "max",
     "median",
     "sum",
     "count",
     "first",
     "last"
    ],
    "title": "Aggregate"
   },
...
   "GroupBy": {
    "type": "string",
    "enum": [
     "entity",
     "device"
    ],
    "title": "GroupBy"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "Layout": {
    "type": "string",
    "enum": [
     "series",
     "long",
     "wide"
    ],
    "title": "Layout"
   },
...
   "PageResponse_MeasurementRow_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/MeasurementRow"
      },
      "type": "array",
      "title": "Items"
     },
     "next_cursor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Cursor"
     }
    },
    "type": "object",
    "required": [
     "items"
    ],
    "title": "PageResponse[MeasurementRow]"
   },
...
   "PageResponse_SavedViewRead_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/SavedViewRead"
      },
      "type": "array",
      "title": "Items"
     },
     "next_cursor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Cursor"
     }
    },
    "type": "object",
    "required": [
     "items"
    ],
    "title": "PageResponse[SavedViewRead]"
   },
...
   "SavedViewCreate": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 200,
      "minLength": 1,
      "title": "Name"
     },
     "view": {
      "additionalProperties": true,
      "type": "object",
      "title": "View"
     },
     "schema_version": {
      "type": "integer",
      "minimum": 1.0,
      "title": "Schema Version",
      "default": 1
     }
    },
    "type": "object",
    "required": [
     "name",
     "view"
    ],
    "title": "SavedViewCreate"
   },
   "SavedViewRead": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "project_id": {
      "type": "string",
      "format": "uuid",
      "title": "Project Id"
     },
     "created_by": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Created By"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "view": {
      "additionalProperties": true,
      "type": "object",
      "title": "View"
     },
     "schema_version": {
      "type": "integer",
      "title": "Schema Version"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "updated_at": {
      "type": "string",
      "format": "date-time",
      "title": "Updated At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "project_id",
     "created_by",
     "name",
     "view",
     "schema_version",
     "created_at",
     "updated_at"
    ],
    "title": "SavedViewRead"
   },
   "SavedViewUpdate": {
    "properties": {
     "name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200,
        "minLength": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "view": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "View"
     },
     "schema_version": {
      "anyOf": [
       {
        "type": "integer",
        "minimum": 1.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Schema Version"
     }
    },
    "type": "object",
    "title": "SavedViewUpdate"
   },
...
   "SeriesResponse": {
    "properties": {
     "time_from": {
      "type": "string",
      "format": "date-time",
      "title": "Time From"
     },
     "time_to": {
      "type": "string",
      "format": "date-time",
      "title": "Time To"
     },
     "bucket": {
      "type": "string",
      "title": "Bucket"
     },
     "bucket_seconds": {
      "type": "integer",
      "title": "Bucket Seconds"
     },
     "automatic_bucket": {
      "type": "boolean",
      "title": "Automatic Bucket"
     },
     "aggregates": {
      "items": {
       "$ref": "#/components/schemas/Aggregate"
      },
      "type": "array",
      "title": "Aggregates"
     },
     "group_by": {
      "$ref": "#/components/schemas/GroupBy"
     },
     "layout": {
      "$ref": "#/components/schemas/Layout"
     },
     "series": {
      "anyOf": [
       {
        "items": {
         "$ref": "#/components/schemas/Series"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Series"
     },
     "columns": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Columns"
     },
     "rows": {
      "anyOf": [
       {
        "items": {
         "additionalProperties": true,
         "type": "object"
        },
        "type": "array"
       },
       {
        "items": {
         "items": {},
         "type": "array"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rows"
     }
    },
    "type": "object",
    "required": [
     "time_from",
     "time_to",
     "bucket",
     "bucket_seconds",
     "automatic_bucket",
     "aggregates",
     "group_by",
     "layout"
    ],
    "title": "SeriesResponse"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "/api/v1/auth/login"
     }
    }
   }
  }
 }
}